Changed orbital_period to a double type

+++ database/migrations/2015_04_19_195553_create_planets_table.php
@@ -22,7 +22,7 @@ class CreatePlanetsTable extends Migration {
             $table->string('type');
             $table->double('gravity');
             $table->integer('distance_from_sun');
-            $table->integer('orbital_period');
+            $table->double('orbital_period');
             $table->boolean('ring');
 			$table->timestamps();
 		});
